requirements.txt
pipwin==0.5.1
PyAudio
pillow==0.4.0
PyAutoGUI==0.9.53
selenium==4.0.0
robotframework==4.1.2
robotframework-seleniumlibrary==5.1.3
requests==2.26.0
pygame==2.1.0
opencv-python==4.5.4.58
moviepy==1.0.3
unidecode==1.3.2
environs==9.4.0